Failed attempt! Moved extra field template in theme-named dir.
This does not work!

I think it is because the template loader
`utils.template_override_middleware.Loader` does not look for
theme-dirs in django apps.

It is probably possible to modify the loader (or to add a new one) to
look for theme-templates in apps template dirs, but it feels that I'm
just adding complexity...