pesky "script_prefix": read the comment
TODO: reverse() uses the `script_prefix` which is set onto the
process's thread by (?) Janeway's middleware to keep track of the
journal (if using a path as opposet to a domain) (?). The prefix is
set by any call to the journal. But if reverse() is called before the
prefix is